Add rating and movie services with repository support
Implemented `RatingService` and `MovieService` to manage movie ratings and details. Added `RatingRepository` for database interactions. Integrated the services and repository in the DI container, ensuring application-wide availability and usage. Updated `Movie` model to use `float` for ratings and adjusted related calculations.

+++ src/CinemaHub.Application/Models/Movie.cs
@@ -35,7 +35,7 @@ public partial class Movie
     }
 
     [NotMapped]
-    public double? Rating { get; set; }
+    public float? Rating { get; set; }
     [NotMapped]
     public int? UserRating { get; set; }
